refactor user controller and validation; enhance error handling and add 404 response for non-existent routes

+++ src/controllers/user.controller.js
@@ -1,20 +1,15 @@
 import {User} from "../models/user.model.js";
 import {apiError} from "../utils/apiError.js";
+import mongoose from "mongoose";
 import {apiResponse} from "../utils/apiResponse.js";
 
 // 1.Creating a new User
 const createUser=async(req,res)=>{
     const {name,email,age}=req.body;
 
-    if([name,email,age].some((field)=>field===undefined))
-    {
-        throw new apiError(400,"All fields are required");
-    }
-
     const existingUser=await User.findOne(
         {email:email}
     )
-    
     if(existingUser){
         throw new apiError(400,"User already exists");
     }
@@ -34,13 +29,15 @@ const getAllUsers=async(req,res)=>{
         return res.status(200).json(new apiResponse(200,"All Users",users));
     }
     catch(err){
-        throw new apiError(400,err.message);
+        throw new apiError(500,err.message);
     }
 };
 
 // 3.Retrieving a single User
 const getOneUser=async(req,res)=>{
     const {id}=req.params;
+
+    // Check if the id is a valid ObjectId
     if(!mongoose.Types.ObjectId.isValid(id)){
         throw new apiError(400,"Invalid User ID"); 
     }
@@ -60,6 +57,20 @@ const getOneUser=async(req,res)=>{
 const updateUser=async(req,res)=>{
     const {id}=req.params;
     const updates=req.body;
+
+    // Check if the id is a valid ObjectId
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        throw new apiError(400,"Invalid User ID"); 
+    }
+
+    if(updates.email){
+        const existingUser=await User.findOne(
+            {email:updates.email}
+        )
+        if(existingUser){
+            throw new apiError(400,"User already exists");
+        }
+    }
 
     try{
         if(Object.keys(req.body).length===0){
@@ -90,6 +101,11 @@ const updateUser=async(req,res)=>{
 // 5.Deleting a User
 const deleteUser=async(req,res)=>{
     const {id}=req.params;
+
+        // Check if the id is a valid ObjectId
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        throw new apiError(400,"Invalid User ID"); 
+    }
     try{
         const user=await User.findByIdAndDelete(id);
         if(!user){
